requirements.txt
streamlit
flask
pyinstaller
requests
python-dotenv